requirements.txt
altgraph==0.17.3
future==0.18.3
pefile==2022.5.30
Pillow==9.4.0
progressbar==2.5
pyinstaller==5.7.0
pyinstaller-hooks-contrib==2022.15
pywin32-ctypes==0.2.0
